# ARROW-6385: [C++] Use xxh3 instead of custom hashing code for non-tiny strings