… cpp/src/arrow/vendored/xxhash.h