This yields better performance in addition to relying on less custom code.