@pitrou 2019-09-03
Benchmarks on Ubuntu 18.04 with gcc 7.4 (x86-64):
* before
```
HashIntegers           12776 ns        12769 ns       154535 bytes_per_second=11.6694G/s items_per_second=1.56623G/s
HashSmallStrings       81587 ns        81541 ns        25652 bytes_per_second=2.49555G/s items_per_second=245.275M/s
HashMediumStrings     483494 ns       483223 ns         4328 bytes_per_second=2.68286G/s items_per_second=41.3888M/s
HashLargeStrings      323605 ns       323421 ns         6577 bytes_per_second=5.92472G/s items_per_second=6.18389M/s

BuildStringDictionaryArray          1462602077 ns   1461951271 ns            1 bytes_per_second=233.616M/s

BuildStringDictionary                 7696631 ns      7686495 ns           91 bytes_per_second=39.2773M/s
UniqueString10bytes/4194304/1024     71341945 ns     71312810 ns           10 bytes_per_second=560.909M/s
UniqueString10bytes/4194304/10240   103873422 ns    103832297 ns            7 bytes_per_second=385.237M/s
UniqueString100bytes/4194304/1024   212966741 ns    212833819 ns            3 bytes_per_second=1.83535G/s
UniqueString100bytes/4194304/10240  274777819 ns    274610854 ns            3 bytes_per_second=1.42247G/s
```
* after
```
HashIntegers           12412 ns        12410 ns       168589 bytes_per_second=12.0078G/s items_per_second=1.61167G/s
HashSmallStrings       59447 ns        59430 ns        35245 bytes_per_second=3.42401G/s items_per_second=336.528M/s
HashMediumStrings     171430 ns       171386 ns        12225 bytes_per_second=7.56435G/s items_per_second=116.696M/s
HashLargeStrings      133742 ns       133705 ns        15833 bytes_per_second=14.3314G/s items_per_second=14.9583M/s

BuildStringDictionaryArray          1525536403 ns   1524966882 ns            1 bytes_per_second=223.962M/s

BuildStringDictionary                 7817772 ns      7804228 ns           84 bytes_per_second=38.6848M/s
UniqueString10bytes/4194304/1024     73274423 ns     73247744 ns           10 bytes_per_second=546.092M/s
UniqueString10bytes/4194304/10240   105917896 ns    105879905 ns            7 bytes_per_second=377.787M/s
UniqueString100bytes/4194304/1024   185390781 ns    185284167 ns            4 bytes_per_second=2.10825G/s
UniqueString100bytes/4194304/10240  253262101 ns    253075255 ns            3 bytes_per_second=1.54351G/s
```