@pitrou 2019-09-03
Benchmarks on Ubuntu 18.04 with clang 7.0 (x86-64):
* before
```
HashIntegers            7818 ns         7817 ns       268959 bytes_per_second=19.0637G/s items_per_second=2.55868G/s
HashSmallStrings      134226 ns       134189 ns        15627 bytes_per_second=1.51645G/s items_per_second=149.044M/s
HashMediumStrings     569914 ns       569758 ns         3680 bytes_per_second=2.27538G/s items_per_second=35.1026M/s
HashLargeStrings      315351 ns       315265 ns         6615 bytes_per_second=6.07801G/s items_per_second=6.34388M/s

BuildStringDictionaryArray          1565437184 ns   1564331162 ns            1 bytes_per_second=218.326M/s

BuildStringDictionary                 7675061 ns      7668470 ns           92 bytes_per_second=39.3696M/s
UniqueString10bytes/4194304/1024     68233039 ns     68205470 ns           10 bytes_per_second=586.463M/s
UniqueString10bytes/4194304/10240   101163657 ns    101119925 ns            7 bytes_per_second=395.57M/s
UniqueString100bytes/4194304/1024   199345875 ns    199220110 ns            4 bytes_per_second=1.96077G/s
UniqueString100bytes/4194304/10240  269394590 ns    269214721 ns            3 bytes_per_second=1.45098G/s
```
* after
```
HashIntegers            7796 ns         7794 ns       269360 bytes_per_second=19.1192G/s items_per_second=2.56614G/s
HashSmallStrings       82722 ns        82701 ns        25383 bytes_per_second=2.46055G/s items_per_second=241.834M/s
HashMediumStrings     179893 ns       179847 ns        11677 bytes_per_second=7.20846G/s items_per_second=111.206M/s
HashLargeStrings      125845 ns       125812 ns        16942 bytes_per_second=15.2306G/s items_per_second=15.8968M/s

BuildStringDictionaryArray          1565018864 ns   1564329811 ns            1 bytes_per_second=218.327M/s

BuildStringDictionary                 7617240 ns      7603062 ns           92 bytes_per_second=39.7083M/s
UniqueString10bytes/4194304/1024     68524774 ns     68497319 ns           10 bytes_per_second=583.964M/s
UniqueString10bytes/4194304/10240   102208886 ns    102174316 ns            7 bytes_per_second=391.488M/s
UniqueString100bytes/4194304/1024   181762167 ns    181652423 ns            4 bytes_per_second=2.1504G/s
UniqueString100bytes/4194304/10240  250831357 ns    250640792 ns            3 bytes_per_second=1.55851G/s
```