@pitrou 2019-09-03
@ursabot benchmark --benchmark-filter=Kernel