@pitrou 2019-09-03
Hmm... the almost identical numbers reported by ursabot seem a bit unlikely, since we're completely changing algorithms here. @fsaintjacques