@pitrou 2019-09-04
> Looks like some MSVC warnings need to be suppressed

Done.